requirements/requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.24.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==10.0.1
pygments==2.13.0
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
stack-data==0.6.2
traitlets==5.8.0
wcwidth==0.2.5
yfinance==0.2.32
python-bcb==0.2.0
